CareercraftAi/CarrierCraftAi/EtMarkTracker/requirements.txt

email-validator>=2.2.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
openai>=1.79.0
psycopg2-binary>=2.9.10
werkzeug>=3.1.3
requests-oauthlib>=2.0.0
